requirements.txt
easyocr
numpy
Pillow
torch
tk
